+++ src/main/java/step2/domain/RandomCarDistanceGenerator.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import java.util.concurrent.ThreadLocalRandom;
