+++ src/main/java/step2/view/InputView.java
@@ -1,4 +1,6 @@
-package step2;
+package step2.view;
+
+import step2.domain.GameRunner;
 
 public interface InputView {
     GameRunner makeGameRunner();
