+++ src/main/java/step2/view/OutputView.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.view;
 
 public interface OutputView {
     void print(final String output);
