+++ src/main/java/step2/view/format/GameResultFormatter.java
@@ -1,4 +1,6 @@
-package step2;
+package step2.view.format;
+
+import step2.domain.GameResult;
 
 public interface GameResultFormatter<T> {
     T format(final GameResult gameResult);
